// src/types/BlockTypes.enum.ts
export enum BlockTypesEnum {
  H1 = "H1",
  H2 = "H2",
  H3 = "H3",
  H4 = "H4",
  H5 = "H5",
  H6 = "H6",
  PARAGRAPH = "PARAGRAPH",
  CHECKBOX = "CHECKBOX",
  QUOTE = "QUOTE",
  CALLOUT = "CALLOUT",
  HR = "HR",
  LIST = "LIST",
  TOGGLE = "TOGGLE",
  TABLE = "TABLE",
  CODE_BLOCK = "CODE_BLOCK",
  IMAGE = "IMAGE",
  VIDEO = "VIDEO",
  AUDIO = "AUDIO",
  HTML = "HTML",
  EMBEDDING = "EMBEDDING",
  LATEX = "LATEX",
}